pyproject.toml
[project]
name = "legal-hallucination-benchmark"
version = "0.1.0"
description = "A benchmark for evaluating LLM tendency to hallucinate legal citations across UK and UAE jurisdictions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pandas>=2.0",
    "openai>=1.0",
    "anthropic>=0.40",
    "google-genai>=1.0",
    "numpy>=1.26",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]
